requirements.txt
asn1crypto==0.24.0
bottle==0.12.13
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
CodeConvert==2.0.5
cryptography==2.2.2
diff-match-patch==20121119
Django==2.0.7
django-admin==1.3.0
django-appconf==1.0.2
django-ckeditor==5.6.1
django-compressor==2.2
django-debug-toolbar==1.9.1
django-excel-response2==2.0.8
django-haystack==2.8.1
django-import-export==1.0.1
django-js-asset==1.1.0
django-models-ext==1.1.6
django-pagedown==1.0.5
django-six==1.0.4
django-suit==0.2.26
django-uuslug==1.1.8
et-xmlfile==1.0.1
furl==1.2
idna==2.7
jdcal==1.4
jieba==0.39
jsonpickle==0.9.6
mistune==0.8.3
odfpy==1.3.6
openpyxl==2.5.4
orderedmultidict==1.0
Pillow==5.2.0
pycparser==2.18
Pygments==2.2.0
PyMySQL==0.9.2
python-dateutil==2.7.3
python-slugify==1.2.5
python3-memcached==1.51
pytz==2018.5
PyYAML==3.13
raven==6.9.0
rcssmin==1.0.6
requests==2.19.1
rjsmin==1.0.12
screen==1.0.1
shortuuid==0.5.0
six==1.11.0
sqlparse==0.2.4
tablib==0.12.1
TimeConvert==1.4.3
tzlocal==1.5.1
unicodecsv==0.14.1
Unidecode==1.0.22
urllib3==1.23
virtualenv==16.0.0
WeRoBot==1.6.0
Whoosh==2.7.4
xlrd==1.1.0
xlwt==1.3.0
xmltodict==0.11.0
